:wrench: Add Font typeface & seperate header comp
