+++ gatsby-browser.js
@@ -1,0 +1,2 @@
+// Load Lato typeface
+require('typeface-lato')
